requirements.txt
streamlit==0.74.1
tensorflow_cpu==2.4.0
scipy==1.6.0
numpy==1.19.5
opencv_python==4.5.1.48
